+++ src/fund/components/FundCardAmountInputTypeSwitch.tsx
@@ -1,11 +1,6 @@
-import { formatFiatAmount } from '@/internal/utils/formatFiatAmount';
-import { truncateDecimalPlaces } from '@/internal/utils/truncateDecimalPlaces';
-import { useCallback, useMemo } from 'react';
-import { Skeleton } from '../../internal/components/Skeleton';
-import { useIcon } from '../../internal/hooks/useIcon';
-import { cn, pressable, text } from '../../styles/theme';
 import type { FundCardAmountInputTypeSwitchPropsReact } from '../types';
 import { useFundContext } from './FundCardProvider';
+import { AmountInputTypeSwitch } from '@/internal/components/amount-input/AmountInputTypeSwitch';
 
 export const FundCardAmountInputTypeSwitch = ({
   className,
@@ -21,59 +16,18 @@ export const FundCardAmountInputTypeSwitch = ({
     currency,
   } = useFundContext();
 
-  const iconSvg = useIcon({ icon: 'toggle' });
-
-  const handleToggle = useCallback(() => {
-    setSelectedInputType(selectedInputType === 'fiat' ? 'crypto' : 'fiat');
-  }, [selectedInputType, setSelectedInputType]);
-
-  const formatCrypto = useCallback(
-    (amount: string) => {
-      return `${truncateDecimalPlaces(amount || '0', 8)} ${asset}`;
-    },
-    [asset],
-  );
-
-  const amountLine = useMemo(() => {
-    return (
-      <span data-testid="ockAmountLine" className={cn(text.label1)}>
-        {selectedInputType === 'fiat'
-          ? formatCrypto(fundAmountCrypto)
-          : formatFiatAmount({
-              amount: fundAmountFiat,
-              currency: currency,
-              minimumFractionDigits: 0,
-            })}
-      </span>
-    );
-  }, [
-    fundAmountCrypto,
-    fundAmountFiat,
-    selectedInputType,
-    formatCrypto,
-    currency,
-  ]);
-
-  if (exchangeRateLoading || !exchangeRate) {
-    return <Skeleton className="h-[1.625rem]" />;
-  }
-
   return (
-    <div className={cn('flex items-center', className)}>
-      <button
-        type="button"
-        aria-label="amount type switch"
-        className={cn(
-          pressable.default,
-          'mr-1 rounded-full p-1 opacity-50 transition-opacity hover:opacity-100',
-        )}
-        data-testid="ockAmountTypeSwitch"
-        onClick={handleToggle}
-      >
-        <div className="h-[1.125rem] w-[1.125rem]">{iconSvg}</div>
-      </button>
-      <div className="w-full truncate">{amountLine}</div>
-    </div>
+    <AmountInputTypeSwitch
+      selectedInputType={selectedInputType}
+      setSelectedInputType={setSelectedInputType}
+      asset={asset}
+      fiatAmount={fundAmountFiat}
+      cryptoAmount={fundAmountCrypto}
+      exchangeRate={exchangeRate}
+      exchangeRateLoading={exchangeRateLoading}
+      currency={currency}
+      className={className}
+    />
   );
 };
